Deleted the turret heading calculation class and added it's functionality to the turret hardware class.
