+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/LOADCode/Main_/Hardware_/Actuators_/Turret.java
@@ -4,18 +4,19 @@ import com.pedropathing.geometry.Pose;
 import com.qualcomm.robotcore.eventloop.opmode.OpMode;
 import com.qualcomm.robotcore.hardware.DcMotor;
 
-import org.firstinspires.ftc.teamcode.LOADCode.Main_.Hardware_.Calculation_.Turret_Heading;
+
 
 import dev.nextftc.control.feedback.PIDCoefficients;
 import dev.nextftc.control.feedforward.BasicFeedforwardParameters;
 
+
 public class Turret {
+
+
     public final Devices.DcMotorExClass rotation = new Devices.DcMotorExClass();
     public final Devices.DcMotorExClass flywheel = new Devices.DcMotorExClass();
     public final Devices.ServoClass hood = new Devices.ServoClass();
     public final Devices.ServoClass gate = new Devices.ServoClass();
-
-    private final Turret_Heading targeting = new Turret_Heading();
 
     public static PIDCoefficients turretCoefficients = new PIDCoefficients(0.002, 0, 0);
     public static PIDCoefficients flywheelCoefficients = new PIDCoefficients(0, 0, 0);
@@ -54,7 +55,7 @@ public class Turret {
      * @param targetRedGoal Set this to true to target the red goal, otherwise targets the blue goal.
      */
     public void updateAimbot(Pose robotPose, boolean targetRedGoal){
-        rotation.setAngle(targeting.calcLocalizer(robotPose, targetRedGoal));
+        rotation.setAngle(calcLocalizer(robotPose, targetRedGoal));
     }
 
     public void setGate(gatestate state){
@@ -63,6 +64,16 @@ public class Turret {
         }else if (state == gatestate.CLOSED){
             gate.setAngle(0.25);
         }
+    }
+
+    public double calcLocalizer (Pose robotPose, boolean targetRedGoal){
+        Pose goalPose = new Pose(0,144,0);
+        if (targetRedGoal) {goalPose = new Pose(144, 144, 0);}
+
+        return Math.toDegrees(Math.atan2(
+                goalPose.getY()-robotPose.getY(),
+                goalPose.getX()-robotPose.getX())
+        ) - Math.toDegrees(robotPose.getHeading()) + 180;
     }
 
     public void setFlywheelRPM(double rpm){
