--- TeamCode/src/main/java/org/firstinspires/ftc/teamcode/LOADCode/Main_/Hardware_/Calculation_/Turret_Heading.java
@@ -1,58 +1,0 @@
-/* Copyright (c) 2022 FIRST. All rights reserved.
- *
- * Redistribution and use in source and binary forms, with or without modification,
- * are permitted (subject to the limitations in the disclaimer below) provided that
- * the following conditions are met:
- *
- * Redistributions of source code must retain the above copyright notice, this list
- * of conditions and the following disclaimer.
- *
- * Redistributions in binary form must reproduce the above copyright notice, this
- * list of conditions and the following disclaimer in the documentation and/or
- * other materials provided with the distribution.
- *
- * Neither the name of FIRST nor the names of its contributors may be used to endorse or
- * promote products derived from this software without specific prior written permission.
- *
- * NO EXPRESS OR IMPLIED LICENSES TO ANY PARTY'S PATENT RIGHTS ARE GRANTED BY THIS
- * LICENSE. THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS
- * "AS IS" AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO,
- * THE IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE
- * ARE DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT OWNER OR CONTRIBUTORS BE LIABLE
- * FOR ANY DIRECT, INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL
- * DAMAGES (INCLUDING, BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR
- * SERVICES; LOSS OF USE, DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER
- * CAUSED AND ON ANY THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY,
- * OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE
- * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
- */
-
-package org.firstinspires.ftc.teamcode.LOADCode.Main_.Hardware_.Calculation_;
-
-import com.pedropathing.geometry.Pose;
-
-public class Turret_Heading {
-    /**
-     * @param robotPose The pose of the robot, gotten from PedroPathing's localization
-     * @param targetRedGoal Set this to true to target the red goal, otherwise targets the blue goal.
-     * @return The angle to set the turret to in degrees.
-     */
-    public double calcLocalizer (Pose robotPose, boolean targetRedGoal){
-        Pose goalPose = new Pose(0,144,0);
-        if (targetRedGoal) {goalPose = new Pose(144, 144, 0);}
-
-        return Math.toDegrees(Math.atan2(
-                goalPose.getY()-robotPose.getY(),
-                goalPose.getX()-robotPose.getX())
-        ) - Math.toDegrees(robotPose.getHeading()) + 180;
-    }
-
-    /**
-     * @param robotPose The pose of the robot, gotten from PedroPathing's localization
-     * @param targetRedGoal Set this to true to target the red goal, otherwise targets the blue goal.
-     * @return The angle to set the turret to in degrees.
-     */
-    public double calcAprilTag (Pose robotPose, boolean targetRedGoal){
-        return 0;
-    }
-}
